[refactor] 관리자 알림메시지 대상 목록 컴포넌트 리팩터링
